Update Art Book Next theme
Changes:
- fixes system logos (atarijaguar, pcenginecd)
- adds/updates system art (auto-neverplayed, auto-lightgun, supergrafx, now-playing, 3do, 3ds, amiga, amigacd32, atari5200, atari7800, intellivision
- adds "now-playing" collection 

Details on "Now Playing"
---
A collection that enables a user to specifically add games they are currently playing.   Its meant to be hand curated and to focus on games a user is looking to finish in their library.

In larger library I look for a way to curate a specific set of games at any given time that I am trying to play through to the end (or at least until I decide I am done with them)
I think of this list differently than 'last played' because as a user I may pick up a random game to try but am not actively going to play it after that.  I also think of this list differently than 'favorites' because I look at that list as a curated list of long time favorites - not necessarily games I am actively playing.