Just refer manual builders to the Dockerfile for the prerequisites.
Don't maintain the same list in two places.

After careful review, the only packages listed in the help doc that
weren't in the Dockerfile were
- ca-certificates
- gnupg
- qemu-user-static

The first two are only needed to install Docker, and thus do not belong
in the manual build section.  The last doesn't appear to be needed for
building, since the Docker container doesn't have it.