Attempt to show FPS for all video plugins in mupen64plus-sa.
These doesn't seem to be enough, at least not in RK3566-X55.

Not sure if these settings still work in upstream mupen, but as long as
they still exist we should probably set them.  If someone wants to debug
why they don't work in JELOS, it's one less thing to fix.