Set "config_save_on_exit" to false by default
This sets config_save_on_exit to false by default in retroarch.cfg for all devices.  Should help reduce the ability for users to get into an odd state in retroarch.  With it on by default there was a high chance of changes a users made in config for a game being saved into the global config.