EXCLUDE_PKGS: fix bugs
skip in install, will skip when marked for installation/inclusion
skip in build, will skip when dependency.

This is a bit icky as I think you could assume in catch22 situation,
Where something is marked to be skipped, but something else (marked for
installtion) depends on it. However the main reason for EXCLUDE_PKGS atm
is to exclude emulators in BASE DEVICES (RK-ARMV8-A) that shall be
device optimized for actual devices. Later we probably want to have
package flags that signal device optimization. EXCLUDE_PKGS is probably
to be marked as a power-user feature to be used at your own risk.